Refactoring
Former-commit-id: 29c6556b3d6516b74f540a22b2f55f2145792f57

+++ 3DCast/src/3DCast/Scene/Components.h
@@ -141,7 +141,7 @@ namespace Cast::Component {
 					UpdateTransformMatrix();
 
 				if (transformRegistry)
-					transformRegistry->AddData(&Transform, sizeof(glm::mat4), bufferPosition);
+					transformRegistry->AddData(&Transform, sizeof(glm::mat4), (int)bufferPosition);
 			}
 		}
 	};
